--- Score/Score/Model/AuthToken.swift
@@ -1,13 +1,0 @@
-//
-//  AuthToken.swift
-//  Score
-//
-//  Created by sole on 5/21/24.
-//
-
-import Foundation
-
-struct AuthToken: Equatable {
-    let accessToken: String
-    let refreshToken: String
-}
