--- Score/Score/Model/SCError.swift
@@ -1,13 +1,0 @@
-//
-//  SCError.swift
-//  Score
-//
-//  Created by sole on 5/21/24.
-//
-
-import Foundation
-import KakaoSDKCommon
-
-struct SCError: SCErrorProtocol {
-    
-}
