+++ Score/Score/Model/User.swift
@@ -10,8 +10,7 @@ import Foundation
 //MARK: - User
 
 struct User: Equatable,
-             Hashable,
-             Codable {
+             Hashable {
     // identifier
     let nickName: String
     let profileImageName: String
@@ -31,26 +30,11 @@ struct User: Equatable,
                                           weight: 50,
                                           schoolName: "감자대학교",
                                           grade: 1)
-    
-//    //MARK: - CodingKeys
-//    
-//    enum CodingKeys: String,
-//                     CodingKey {
-//        case nickname
-//        case profileImageName
-//        case gender
-//        case height
-//        case weight
-//        case schoolName
-//        case grade
-//    }
 }
 
 //MARK: - Gender
 
-enum Gender: String,
-             CaseIterable,
-             Codable {
+enum Gender: String, CaseIterable {
     static let sex: [Gender] = [.male,
                                 .female]
