--- Score/Score/Protocol/SCErrorProtocol.swift
@@ -1,12 +1,0 @@
-//
-//  SCErrorProtocol.swift
-//  Score
-//
-//  Created by sole on 5/21/24.
-//
-
-import Foundation
-
-protocol SCErrorProtocol: Error {
-    
-}
