--- Score/Score/Reducer/App/Auth/KeyChainStore.swift
@@ -1,30 +1,0 @@
-//
-//  KeyChainStore.swift
-//  Score
-//
-//  Created by sole on 5/21/24.
-//
-
-import Foundation
-
-final class KeyChainStore {
-    static let shared: KeyChainStore = .init()
-    
-    private init() {}
-    
-    func create() {
-        
-    }
-    
-    func read() {
-        
-    }
-    
-    func update() {
-        
-    }
-    
-    func delete() {
-        
-    }
-}
