+++ Score/Score/Resource/Assets.xcassets/Auth/kakao.imageset/Contents.json
@@ -1,17 +1,17 @@
 {
   "images" : [
     {
-      "filename" : "kakao.png",
+      "filename" : "카카오 로고.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
-      "filename" : "kakao@2x.png",
+      "filename" : "카카오 로고@2x.png",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
-      "filename" : "kakao@3x.png",
+      "filename" : "카카오 로고@3x.png",
       "idiom" : "universal",
       "scale" : "3x"
     }
