+++ Score/Score/Resource/Assets.xcassets/Auth/naver.imageset/Contents.json
@@ -1,17 +1,17 @@
 {
   "images" : [
     {
-      "filename" : "naver.png",
+      "filename" : "네이버 로고.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
     {
-      "filename" : "naver@2x.png",
+      "filename" : "네이버 로고@2x.png",
       "idiom" : "universal",
       "scale" : "2x"
     },
     {
-      "filename" : "naver@3x.png",
+      "filename" : "네이버 로고@3x.png",
       "idiom" : "universal",
       "scale" : "3x"
     }
