+++ Score/Score/Resource/Assets.xcassets/Auth/naver.imageset/네이버 로고@2x.png
