+++ Score/Score/Resource/Assets.xcassets/brandIcon/brandIcon.imageset/대지 1@4x 4@3x.png
