+++ Score/Score/Resource/Info.plist
@@ -2,24 +2,12 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>CFBundleURLTypes</key>
-	<array>
-		<dict>
-			<key>CFBundleTypeRole</key>
-			<string>Editor</string>
-			<key>CFBundleURLSchemes</key>
-			<array>
-				<string>kakao</string>
-				<string>9f1cd396bd1a8ab368b3cf4ddf2ab75f</string>
-			</array>
-		</dict>
-	</array>
-	<key>KAKAO_APP_ID</key>
-	<string>$(KAKAO_APP_ID)</string>
 	<key>NAVER_CLIENT_ID</key>
 	<string>$(NAVER_CLIENT_ID)</string>
 	<key>NAVER_SECRET_KEY</key>
 	<string>$(NAVER_SECRET_KEY)</string>
+	<key>KAKAO_APP_ID</key>
+	<string>$(KAKAO_APP_ID)</string>
 	<key>UIAppFonts</key>
 	<array>
 		<string>PretendardVariable.ttf</string>
